+++ src/test/jest.ts
@@ -4,21 +4,30 @@ import { Address, Cell, Slice } from '@ton/core';
 import { FlatTransactionComparable, compareTransactionForTest } from './transaction';
 import { CompareResult } from './interface';
 import { compareAddressForTest, compareCellForTest, compareSliceForTest } from './comparisons';
+import { compareThrownExitCodeForTest } from './exitCode';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function wrapComparer<T>(comparer: (subject: any, cmp: T) => CompareResult): MatcherFunction<[cmp: T]> {
+function extractResult(result: CompareResult) {
+    return {
+        pass: result.pass,
+        message: () => {
+            if (result.pass) {
+                return result.negMessage();
+            } else {
+                return result.posMessage();
+            }
+        },
+    };
+}
+
+function wrapComparer<T>(
+    comparer: (subject: unknown, cmp: T) => CompareResult | Promise<CompareResult>,
+): MatcherFunction<[cmp: T]> {
     return function (actual, cmp) {
         const result = comparer(actual, cmp);
-        return {
-            pass: result.pass,
-            message: () => {
-                if (result.pass) {
-                    return result.negMessage();
-                } else {
-                    return result.posMessage();
-                }
-            },
-        };
+        if (result instanceof Promise) {
+            return result.then(extractResult);
+        }
+        return extractResult(result);
     };
 }
 
@@ -26,6 +35,7 @@ const toHaveTransaction = wrapComparer(compareTransactionForTest);
 const toEqualCell = wrapComparer(compareCellForTest);
 const toEqualAddress = wrapComparer(compareAddressForTest);
 const toEqualSlice = wrapComparer(compareSliceForTest);
+const toThrowExitCode = wrapComparer(compareThrownExitCodeForTest);
 
 try {
     // eslint-disable-next-line @typescript-eslint/no-require-imports
@@ -37,6 +47,7 @@ try {
             toEqualCell,
             toEqualAddress,
             toEqualSlice,
+            toThrowExitCode,
         });
     // eslint-disable-next-line no-empty
 } catch (_) {}
@@ -49,6 +60,12 @@ declare global {
             toEqualCell(cell: Cell): R;
             toEqualAddress(address: Address): R;
             toEqualSlice(slice: Slice): R;
+
+            /**
+             * @example
+             * await expect(() => blockchain.runGetMethod(contract.address, 'test')).toThrowExitCode(11);
+             */
+            toThrowExitCode(exitCode: number): Promise<R>;
         }
     }
 }