+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `toThrowExitCode` jest matcher and `throwExitCode` chai assertion to validate the `exitCode` of errors thrown from get method calls or external message invocations.
+
 ## [0.7.0] - 2025-06-02
 
 - Add address, opcode and errors description in `toHaveTransaction` method