+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- **feat**: Add support for the Bun.js test runner.
+
+### Changed
+- **docs**: Updated `README.md` with examples and instructions for using the Bun matchers.
+
 ## [0.7.0] - 2025-06-02
 
 - Add address, opcode and errors description in `toHaveTransaction` method